AI/requirements.txt
# Flask Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
mysql-connector-python==8.2.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Optional: Machine Learning (for future Kaggle models)
# numpy==1.26.2
# pandas==2.1.4
# scikit-learn==1.3.2
# joblib==1.3.2

